app/requirements.txt
flask
mysql-connector-python
flask-cors
psycopg2-binary
requests
paramiko